Added reorder_name_key function
REST Commands sent to the DataPower REST Mgmt interface endpoint
/mgmt/actionqueue will fail if the 'name' key is not the first and
the python version is 2.7 or 3.5 as these versions do not perserve
insertion order.

This function is added to ensure that the key 'name' is always the
first key in a dictionary.  This was added to httpapi/rest_mgmt.py
as the connection methods will serialize/deserialize the json data,
effectively undoing any sorting done within the modules.